Move write_file into own module (which is properly GPL licensed, as the original code is).
