+++ plugins/module_utils/sops.py
@@ -1,5 +1,4 @@
 # Copyright (c), Edoardo Tenani <e.tenani@arduino.cc>, 2018-2020
-# Copyright (c), Yanis Guenane <yanis+ansible@guenane.org>, 2016
 # Simplified BSD License (see licenses/simplified_bsd.txt or https://opensource.org/licenses/BSD-2-Clause)
 
 from __future__ import absolute_import, division, print_function
@@ -9,9 +8,6 @@ __metaclass__ = type
 from ansible.module_utils._text import to_text, to_native
 
 from subprocess import Popen, PIPE
-
-import os
-import tempfile
 
 
 # From https://github.com/mozilla/sops/blob/master/cmd/sops/codes/codes.go
@@ -110,44 +106,3 @@ class Sops():
             raise SopsError('to stdout', exit_code, err, decryption=False)
 
         return output
-
-
-def write_file(module, content):
-    '''
-    Writes content into destination file as securely as possible.
-    Uses file arguments from module.
-    '''
-    # Find out parameters for file
-    file_args = module.load_file_common_arguments(module.params)
-    # Create tempfile name
-    tmp_fd, tmp_name = tempfile.mkstemp(prefix=b'.ansible_tmp')
-    try:
-        os.close(tmp_fd)
-    except Exception:
-        pass
-    module.add_cleanup_file(tmp_name)  # if we fail, let Ansible try to remove the file
-    try:
-        try:
-            # Create tempfile
-            file = os.open(tmp_name, os.O_WRONLY | os.O_CREAT | os.O_TRUNC, 0o600)
-            os.write(file, content)
-            os.close(file)
-        except Exception as e:
-            try:
-                os.remove(tmp_name)
-            except Exception:
-                pass
-            module.fail_json(msg='Error while writing result into temporary file: {0}'.format(e))
-        # Update destination to wanted permissions
-        if os.path.exists(file_args['path']):
-            module.set_fs_attributes_if_different(file_args, False)
-        # Move tempfile to final destination
-        module.atomic_move(tmp_name, file_args['path'])
-        # Try to update permissions again
-        module.set_fs_attributes_if_different(file_args, False)
-    except Exception as e:
-        try:
-            os.remove(tmp_name)
-        except Exception:
-            pass
-        module.fail_json(msg='Error while writing result: {0}'.format(e))
