+++ plugins/modules/sops_encrypt.py
@@ -85,7 +85,8 @@ import traceback
 from ansible.module_utils.basic import AnsibleModule, missing_required_lib
 from ansible.module_utils._text import to_text
 
-from ansible_collections.community.sops.plugins.module_utils.sops import Sops, SopsError, write_file
+from ansible_collections.community.sops.plugins.module_utils.io import write_file
+from ansible_collections.community.sops.plugins.module_utils.sops import Sops, SopsError
 
 try:
     import yaml
